requirements.txt


aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-user-agents==0.4.0
frozenlist==1.3.1
geoip2==4.6.0
gunicorn==20.1.0
idna==3.4
maxminddb==2.2.0
multidict==6.0.2
psycopg2-binary==2.9.3
requests==2.28.1
sqlparse==0.4.3
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
yarl==1.8.1

selenium==4.5.0
django-import-export==3.0.1
django-admin-rangefilter==0.9.0
